+++ scripts/sound-processor/consts.js
@@ -1,0 +1,6 @@
+const RAW_FOLDER = 'raw';
+const PROCESSED_FOLDER = 'processed';
+const SOUNDS_FOLDER = 'sounds';
+const TEMP_FOLDER = '_00_temp_00_';
+
+export { RAW_FOLDER, PROCESSED_FOLDER, SOUNDS_FOLDER, TEMP_FOLDER };